remove ParticlesPurelyLongitudinal
