+++ xpart/particles.py
@@ -6,13 +6,12 @@
 import numpy as np
 import xobjects as xo
 
-from ..general import _pkg_root
-from ..constants import PROTON_MASS_EV
-from ..pdg import get_pdg_id_from_name
+from xpart.general import _pkg_root
+from xpart.constants import PROTON_MASS_EV
+from xpart.pdg import get_pdg_id_from_name, get_properties_from_pdg_id, get_mass_from_pdg_id
 
 from scipy.constants import e as qe
 from scipy.constants import c as clight
-from scipy.constants import m_p
 from scipy.constants import epsilon_0
 
 from xobjects import BypassLinked
@@ -21,7 +20,7 @@ from xobjects import BypassLinked
 LAST_INVALID_STATE = -999999999
 
 
-class ParticlesBase(xo.HybridClass):
+class Particles(xo.HybridClass):
     _cname = 'ParticlesData'
 
     size_vars = (
@@ -53,6 +52,10 @@ class ParticlesBase(xo.HybridClass):
             (xo.Float64, 'beta0'),
             (xo.Float64, 's'),
             (xo.Float64, 'zeta'),
+            (xo.Float64, 'x'),
+            (xo.Float64, 'y'),
+            (xo.Float64, 'px'),
+            (xo.Float64, 'py'),
         )
         + part_energy_vars +
         (
@@ -91,6 +94,15 @@ class ParticlesBase(xo.HybridClass):
         'p0c': '_p0c',
         'gamma0': '_gamma0',
         'beta0': '_beta0',
+    }
+
+    _kernels = {
+        'Particles_initialize_rand_gen': xo.Kernel(
+            args=[
+                xo.Arg(xo.ThisClass, name='particles'),
+                xo.Arg(xo.UInt32, pointer=True, name='seeds'),
+                xo.Arg(xo.Int32, name='n_init')],
+            n_threads='n_init')
     }
 
     def __init__(
@@ -181,15 +193,6 @@ class ParticlesBase(xo.HybridClass):
             Identifier of the parent particle (secondary production processes)
 
         """
-
-        if type(self) is ParticlesBase:
-            raise NotImplementedError(
-                'ParticlesBase is an abstract class to be used as a template '
-                'for usable concrete implementations. It serves only a purpose of '
-                'defining the bare minimum C-API. Therefore, it cannot be '
-                'instantiated.'
-            )
-
         if '_xobject' in kwargs.keys():
             # Initialize xobject
             self.xoinitialize(**kwargs)
@@ -1004,7 +1007,7 @@ class ParticlesBase(xo.HybridClass):
 
         context = self._buffer.context
         seeds_dev = context.nparray_to_context_array(seeds)
-        kernel = context.kernels['Particles_initialize_rand_gen', (self._XoStruct,)]
+        kernel = context.kernels['Particles_initialize_rand_gen']
         kernel(particles=self, seeds=seeds_dev, n_init=self._capacity)
 
     def hide_first_n_particles(self, num_particles):
@@ -1304,20 +1307,20 @@ class ParticlesBase(xo.HybridClass):
         # Get io buffer
         src_lines = []
         src_lines.append('''
-        /*gpufun*/
-        /*gpuglmem*/ int8_t* LocalParticle_get_io_buffer(LocalParticle* part){
-            return part->io_buffer;
-        }
+            /*gpufun*/
+            /*gpuglmem*/ int8_t* LocalParticle_get_io_buffer(LocalParticle* part){
+                return part->io_buffer;
+            }
 
-        ''')
+            ''')
 
         # Particles_to_LocalParticle
         src_lines.append('''
-        /*gpufun*/
-        void Particles_to_LocalParticle(ParticlesData source,
-                                        LocalParticle* dest,
-                                        int64_t id,
-                                        int64_t eid){''')
+            /*gpufun*/
+            void Particles_to_LocalParticle(ParticlesData source,
+                                            LocalParticle* dest,
+                                            int64_t id,
+                                            int64_t eid){''')
         for _, vv in cls.size_vars + cls.scalar_vars:
             src_lines.append(
                 f'  dest->{vv} = ParticlesData_get_' + vv + '(source);')
@@ -1334,12 +1337,12 @@ class ParticlesBase(xo.HybridClass):
         # LocalParticle_to_Particles
         src_lines = []
         src_lines.append('''
-        /*gpufun*/
-        void LocalParticle_to_Particles(
-                                        LocalParticle* source,
-                                        ParticlesData dest,
-                                        int64_t id,
-                                        int64_t set_scalar){''')
+            /*gpufun*/
+            void LocalParticle_to_Particles(
+                                            LocalParticle* source,
+                                            ParticlesData dest,
+                                            int64_t id,
+                                            int64_t set_scalar){''')
         src_lines.append('if (set_scalar){')
         for _, vv in cls.size_vars + cls.scalar_vars:
             src_lines.append(
@@ -1358,8 +1361,8 @@ class ParticlesBase(xo.HybridClass):
         src_lines = []
         for tt, vv in cls.per_particle_vars:
             src_lines.append('''
-        /*gpufun*/
-        void LocalParticle_add_to_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
+            /*gpufun*/
+            void LocalParticle_add_to_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
                              + '{')
             src_lines.append(f'#ifndef FREEZE_VAR_{vv}')
             src_lines.append(f'  part->{vv}[part->ipart] += value;')
@@ -1371,8 +1374,8 @@ class ParticlesBase(xo.HybridClass):
         src_lines = []
         for tt, vv in cls.per_particle_vars:
             src_lines.append('''
-        /*gpufun*/
-        void LocalParticle_scale_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
+            /*gpufun*/
+            void LocalParticle_scale_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
                              + '{')
             src_lines.append(f'#ifndef FREEZE_VAR_{vv}')
             src_lines.append(f'  part->{vv}[part->ipart] *= value;')
@@ -1384,8 +1387,8 @@ class ParticlesBase(xo.HybridClass):
         src_lines = []
         for tt, vv in cls.per_particle_vars:
             src_lines.append('''
-        /*gpufun*/
-        void LocalParticle_set_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
+            /*gpufun*/
+            void LocalParticle_set_''' + vv + f'(LocalParticle* part, {tt._c_type} value)'
                              + '{')
             src_lines.append(f'#ifndef FREEZE_VAR_{vv}')
             src_lines.append(f'  part->{vv}[part->ipart] = value;')
@@ -1417,12 +1420,6 @@ class ParticlesBase(xo.HybridClass):
         # Angles
         src_angles_lines = []
         for exact in ['', 'exact_']:
-
-            if ('px' not in [nn for _, nn in cls.per_particle_vars]
-                or 'py' not in [nn for _, nn in cls.per_particle_vars]):
-                # The variable in not in the per_particle_vars
-                continue
-
             # xp (py as transverse) and vice versa
             for xx, yy in [['x', 'y'], ['y', 'x']]:
                 # Getter
@@ -1432,7 +1429,8 @@ class ParticlesBase(xo.HybridClass):
                 if exact == 'exact_':
                     src_angles_lines.append(f'    double const p{yy} = LocalParticle_get_p{yy}(part);')
                     src_angles_lines.append(f'    double const one_plus_delta = 1. + LocalParticle_get_delta(part);')
-                    src_angles_lines.append(f'    double const rpp = 1./sqrt(one_plus_delta*one_plus_delta - px*px - py*py);')
+                    src_angles_lines.append(
+                        f'    double const rpp = 1./sqrt(one_plus_delta*one_plus_delta - px*px - py*py);')
                 else:
                     src_angles_lines.append(f'    double const rpp = LocalParticle_get_rpp(part);')
                 src_angles_lines.append(f'    // INFO: this is not the angle, but sin(angle)')
@@ -1447,7 +1445,8 @@ class ParticlesBase(xo.HybridClass):
                 src_angles_lines.append(f'#ifndef FREEZE_VAR_p{xx}')
                 src_angles_lines.append(f'    double rpp = LocalParticle_get_rpp(part);')
                 if exact == 'exact_':
-                    src_angles_lines.append(f'    // Careful! If {yy}p also changes, use LocalParticle_set_{exact}xp_yp!')
+                    src_angles_lines.append(
+                        f'    // Careful! If {yy}p also changes, use LocalParticle_set_{exact}xp_yp!')
                     src_angles_lines.append(f'    double const {yy}p = LocalParticle_get_{exact}{yy}p(part);')
                     src_angles_lines.append(f'    rpp *= sqrt(1 + xp*xp + yp*yp);')
                 src_angles_lines.append(f'    // INFO: {xx}p is not the angle, but sin(angle)')
@@ -1462,7 +1461,7 @@ class ParticlesBase(xo.HybridClass):
                 src_angles_lines.append(f'void LocalParticle_add_to_{exact}{xx}p(LocalParticle* part, double {xx}p){{')
                 src_angles_lines.append(f'#ifndef FREEZE_VAR_p{xx}')
                 src_angles_lines.append(f'    LocalParticle_set_{exact}{xx}p(part, '
-                             + f'LocalParticle_get_{exact}{xx}p(part) + {xx}p);')
+                                        + f'LocalParticle_get_{exact}{xx}p(part) + {xx}p);')
                 src_angles_lines.append(f'#endif')
                 src_angles_lines.append('}')
                 src_angles_lines.append('')
@@ -1471,7 +1470,7 @@ class ParticlesBase(xo.HybridClass):
                 src_angles_lines.append(f'void LocalParticle_scale_{exact}{xx}p(LocalParticle* part, double value){{')
                 src_angles_lines.append(f'#ifndef FREEZE_VAR_p{xx}')
                 src_angles_lines.append(f'    LocalParticle_set_{exact}{xx}p(part, '
-                             + f'LocalParticle_get_{exact}{xx}p(part) * value);')
+                                        + f'LocalParticle_get_{exact}{xx}p(part) * value);')
                 src_angles_lines.append(f'#endif')
                 src_angles_lines.append('}')
                 src_angles_lines.append('')
@@ -1488,25 +1487,27 @@ class ParticlesBase(xo.HybridClass):
             src_angles_lines.append('}')
             src_angles_lines.append('')
             src_angles_lines.append('/*gpufun*/')
-            src_angles_lines.append(f'void LocalParticle_add_to_{exact}xp_yp(LocalParticle* part, double xp, double yp){{')
+            src_angles_lines.append(
+                f'void LocalParticle_add_to_{exact}xp_yp(LocalParticle* part, double xp, double yp){{')
             src_angles_lines.append(f'    LocalParticle_set_{exact}xp_yp(part, '
-                         + f'LocalParticle_get_{exact}xp(part) + xp, '
-                         + f'LocalParticle_get_{exact}yp(part) + yp);')
+                                    + f'LocalParticle_get_{exact}xp(part) + xp, '
+                                    + f'LocalParticle_get_{exact}yp(part) + yp);')
             src_angles_lines.append('}')
             src_angles_lines.append('')
             src_angles_lines.append('/*gpufun*/')
-            src_angles_lines.append(f'void LocalParticle_scale_{exact}xp_yp(LocalParticle* part, double value_x, double value_y){{')
+            src_angles_lines.append(
+                f'void LocalParticle_scale_{exact}xp_yp(LocalParticle* part, double value_x, double value_y){{')
             src_angles_lines.append(f'    LocalParticle_set_{exact}xp_yp(part, '
-                         + f'LocalParticle_get_{exact}xp(part) * value_x, '
-                         + f'LocalParticle_get_{exact}yp(part) * value_y);')
+                                    + f'LocalParticle_get_{exact}xp(part) * value_x, '
+                                    + f'LocalParticle_get_{exact}yp(part) * value_y);')
             src_angles_lines.append('}')
         src_angles = '\n'.join(src_angles_lines)
 
         # Particle exchangers
         src_exchange = '''
-    /*gpufun*/
-    void LocalParticle_exchange(LocalParticle* part, int64_t i1, int64_t i2){
-    '''
+        /*gpufun*/
+        void LocalParticle_exchange(LocalParticle* part, int64_t i1, int64_t i2){
+        '''
         for tt, vv in cls.per_particle_vars:
             src_exchange += '\n'.join([
                 '\n    {',
@@ -1517,208 +1518,290 @@ class ParticlesBase(xo.HybridClass):
         src_exchange += '}\n'
 
         custom_source = '''
-    /*gpufun*/
-    double LocalParticle_get_energy0(LocalParticle* part){
+        /*gpufun*/
+        double LocalParticle_get_energy0(LocalParticle* part){
 
-        double const p0c = LocalParticle_get_p0c(part);
-        double const m0  = LocalParticle_get_mass0(part);
+            double const p0c = LocalParticle_get_p0c(part);
+            double const m0  = LocalParticle_get_mass0(part);
 
-        return sqrt( p0c * p0c + m0 * m0 );
-    }
-
-    /*gpufun*/
-    void LocalParticle_update_ptau(LocalParticle* part, double new_ptau_value){
-
-        double const beta0 = LocalParticle_get_beta0(part);
-
-        double const ptau = new_ptau_value;
-
-        double const irpp = sqrt(ptau*ptau + 2*ptau/beta0 +1);
-
-        double const new_rpp = 1./irpp;
-        LocalParticle_set_delta(part, irpp - 1.);
-
-        double const new_rvv = irpp/(1 + beta0*ptau);
-        LocalParticle_set_rvv(part, new_rvv);
-        LocalParticle_set_ptau(part, ptau);
-
-        LocalParticle_set_rpp(part, new_rpp );
-    }
-
-    /*gpufun*/
-    void LocalParticle_update_delta(LocalParticle* part, double new_delta_value){
-        double const beta0 = LocalParticle_get_beta0(part);
-        double const delta_beta0 = new_delta_value * beta0;
-        double const ptau_beta0  = sqrt( delta_beta0 * delta_beta0 +
-                                    2. * delta_beta0 * beta0 + 1. ) - 1.;
-
-        double const one_plus_delta = 1. + new_delta_value;
-        double const rvv    = ( one_plus_delta ) / ( 1. + ptau_beta0 );
-        double const rpp    = 1. / one_plus_delta;
-        double const ptau = ptau_beta0 / beta0;
-
-        LocalParticle_set_delta(part, new_delta_value);
-
-        LocalParticle_set_rvv(part, rvv );
-        LocalParticle_set_rpp(part, rpp );
-        LocalParticle_set_ptau(part, ptau );
-
-    }
-
-    /*gpufun*/
-    double LocalParticle_get_pzeta(LocalParticle* part){
-
-        double const ptau = LocalParticle_get_ptau(part);
-        double const beta0 = LocalParticle_get_beta0(part);
-
-        return ptau/beta0;
-
-    }
-
-    /*gpufun*/
-    void LocalParticle_update_pzeta(LocalParticle* part, double new_pzeta_value){
-
-        double const beta0 = LocalParticle_get_beta0(part);
-        LocalParticle_update_ptau(part, beta0*new_pzeta_value);
-
-    }
-
-    /*gpufun*/
-    void increment_at_element(LocalParticle* part0, int64_t const increment){
-
-        //start_per_particle_block (part0->part)
-            LocalParticle_add_to_at_element(part, increment);
-        //end_per_particle_block
-
-    }
-
-    /*gpufun*/
-    void increment_at_turn(LocalParticle* part0, int flag_reset_s){
-
-        //start_per_particle_block (part0->part)
-        LocalParticle_add_to_at_turn(part, 1);
-        LocalParticle_set_at_element(part, 0);
-        if (flag_reset_s>0){
-            LocalParticle_set_s(part, 0.);
+            return sqrt( p0c * p0c + m0 * m0 );
         }
-        //end_per_particle_block
-    }
 
-    /*gpufun*/
-    void increment_at_turn_backtrack(LocalParticle* part0, int flag_reset_s,
-                                     double const line_length,
-                                     int64_t const num_elements){
+        /*gpufun*/
+        void LocalParticle_update_ptau(LocalParticle* part, double new_ptau_value){
 
-        //start_per_particle_block (part0->part)
-        LocalParticle_add_to_at_turn(part, -1);
-        LocalParticle_set_at_element(part, num_elements);
-        if (flag_reset_s>0){
-            LocalParticle_set_s(part, line_length);
+            double const beta0 = LocalParticle_get_beta0(part);
+
+            double const ptau = new_ptau_value;
+
+            double const irpp = sqrt(ptau*ptau + 2*ptau/beta0 +1);
+
+            double const new_rpp = 1./irpp;
+            LocalParticle_set_delta(part, irpp - 1.);
+
+            double const new_rvv = irpp/(1 + beta0*ptau);
+            LocalParticle_set_rvv(part, new_rvv);
+            LocalParticle_set_ptau(part, ptau);
+
+            LocalParticle_set_rpp(part, new_rpp );
         }
-        //end_per_particle_block
-    }
 
-    // check_is_active has different implementation on CPU and GPU
+        /*gpufun*/
+        void LocalParticle_update_delta(LocalParticle* part, double new_delta_value){
+            double const beta0 = LocalParticle_get_beta0(part);
+            double const delta_beta0 = new_delta_value * beta0;
+            double const ptau_beta0  = sqrt( delta_beta0 * delta_beta0 +
+                                        2. * delta_beta0 * beta0 + 1. ) - 1.;
 
-    #define CPU_SERIAL_IMPLEM //only_for_context cpu_serial
-    #define CPU_OMP_IMPLEM //only_for_context cpu_openmp
+            double const one_plus_delta = 1. + new_delta_value;
+            double const rvv    = ( one_plus_delta ) / ( 1. + ptau_beta0 );
+            double const rpp    = 1. / one_plus_delta;
+            double const ptau = ptau_beta0 / beta0;
 
-    #ifdef CPU_SERIAL_IMPLEM
+            LocalParticle_set_delta(part, new_delta_value);
 
-    /*gpufun*/
-    int64_t check_is_active(LocalParticle* part) {
-        int64_t ipart=0;
-        while (ipart < part->_num_active_particles){
-            if (part->state[ipart]<1){
-                LocalParticle_exchange(
-                    part, ipart, part->_num_active_particles-1);
-                part->_num_active_particles--;
-                part->_num_lost_particles++;
+            LocalParticle_set_rvv(part, rvv );
+            LocalParticle_set_rpp(part, rpp );
+            LocalParticle_set_ptau(part, ptau );
+
+        }
+
+        /*gpufun*/
+        double LocalParticle_get_pzeta(LocalParticle* part){
+
+            double const ptau = LocalParticle_get_ptau(part);
+            double const beta0 = LocalParticle_get_beta0(part);
+
+            return ptau/beta0;
+
+        }
+
+        /*gpufun*/
+        void LocalParticle_update_pzeta(LocalParticle* part, double new_pzeta_value){
+
+            double const beta0 = LocalParticle_get_beta0(part);
+            LocalParticle_update_ptau(part, beta0*new_pzeta_value);
+
+        }
+
+        /*gpufun*/
+        void increment_at_element(LocalParticle* part0, int64_t const increment){
+
+            //start_per_particle_block (part0->part)
+                LocalParticle_add_to_at_element(part, increment);
+            //end_per_particle_block
+
+        }
+
+        /*gpufun*/
+        void increment_at_turn(LocalParticle* part0, int flag_reset_s){
+
+            //start_per_particle_block (part0->part)
+            LocalParticle_add_to_at_turn(part, 1);
+            LocalParticle_set_at_element(part, 0);
+            if (flag_reset_s>0){
+                LocalParticle_set_s(part, 0.);
             }
-            else{
-                ipart++;
-            }
+            //end_per_particle_block
         }
 
-        if (part->_num_active_particles==0){
-            return 0;//All particles lost
-        } else {
-            return 1; //Some stable particles are still present
-        }
-    }
+        /*gpufun*/
+        void increment_at_turn_backtrack(LocalParticle* part0, int flag_reset_s,
+                                         double const line_length,
+                                         int64_t const num_elements){
 
-    #else // not CPU_SERIAL_IMPLEM
-    #ifdef CPU_OMP_IMPLEM
-    
-    /*gpufun*/
-    int64_t check_is_active(LocalParticle* part) {
-    #ifndef SKIP_SWAPS
-        int64_t ipart = part->ipart;
-        int64_t endpart = part->endpart;
-        
-        int64_t left = ipart;
-        int64_t right = endpart - 1;
-        int64_t swap_made = 0;
-        int64_t has_alive = 0;
-        
-        if (left == right) return part->state[left] > 0;
-        
-        while (left < right) {
-            if (part->state[left] > 0) {
-                left++;
-                has_alive = 1;
+            //start_per_particle_block (part0->part)
+            LocalParticle_add_to_at_turn(part, -1);
+            LocalParticle_set_at_element(part, num_elements);
+            if (flag_reset_s>0){
+                LocalParticle_set_s(part, line_length);
             }
-            else if (part->state[right] <= 0) right--;
-            else {
-                LocalParticle_exchange(part, left, right);
-                left++;
-                right--;
-                swap_made = 1;
+            //end_per_particle_block
+        }
+
+        // check_is_active has different implementation on CPU and GPU
+
+        #define CPU_SERIAL_IMPLEM //only_for_context cpu_serial
+        #define CPU_OMP_IMPLEM //only_for_context cpu_openmp
+
+        #ifdef CPU_SERIAL_IMPLEM
+
+        /*gpufun*/
+        int64_t check_is_active(LocalParticle* part) {
+            int64_t ipart=0;
+            while (ipart < part->_num_active_particles){
+                if (part->state[ipart]<1){
+                    LocalParticle_exchange(
+                        part, ipart, part->_num_active_particles-1);
+                    part->_num_active_particles--;
+                    part->_num_lost_particles++;
+                }
+                else{
+                    ipart++;
+                }
+            }
+
+            if (part->_num_active_particles==0){
+                return 0;//All particles lost
+            } else {
+                return 1; //Some stable particles are still present
             }
         }
 
-        return swap_made || has_alive;
-    #else
-        return 1;
-    #endif
-    }
-    
-    /*gpufun*/
-    void count_reorganized_particles(LocalParticle* part) {
-        int64_t num_active = 0;
-        int64_t num_lost = 0;
-        
-        for (int64_t i = part->ipart; i < part->endpart; i++) {
-            if (part->state[i] <= -999999999) break;
-            else if (part->state[i] > 0) num_active++;
-            else num_lost++;
+        #else // not CPU_SERIAL_IMPLEM
+        #ifdef CPU_OMP_IMPLEM
+
+        /*gpufun*/
+        int64_t check_is_active(LocalParticle* part) {
+        #ifndef SKIP_SWAPS
+            int64_t ipart = part->ipart;
+            int64_t endpart = part->endpart;
+
+            int64_t left = ipart;
+            int64_t right = endpart - 1;
+            int64_t swap_made = 0;
+            int64_t has_alive = 0;
+
+            if (left == right) return part->state[left] > 0;
+
+            while (left < right) {
+                if (part->state[left] > 0) {
+                    left++;
+                    has_alive = 1;
+                }
+                else if (part->state[right] <= 0) right--;
+                else {
+                    LocalParticle_exchange(part, left, right);
+                    left++;
+                    right--;
+                    swap_made = 1;
+                }
+            }
+
+            return swap_made || has_alive;
+        #else
+            return 1;
+        #endif
         }
-        
-        part->_num_active_particles = num_active;
-        part->_num_lost_particles = num_lost;
-    }
-    
-    #else // not CPU_SERIAL_IMPLEM and not CPU_OMP_IMPLEM
 
-    /*gpufun*/
-    int64_t check_is_active(LocalParticle* part) {
-        return LocalParticle_get_state(part)>0;
-    };
+        /*gpufun*/
+        void count_reorganized_particles(LocalParticle* part) {
+            int64_t num_active = 0;
+            int64_t num_lost = 0;
 
-    #endif // CPU_OMP_IMPLEM
-    #endif // CPU_SERIAL_IMPLEM
+            for (int64_t i = part->ipart; i < part->endpart; i++) {
+                if (part->state[i] <= -999999999) break;
+                else if (part->state[i] > 0) num_active++;
+                else num_lost++;
+            }
 
-    #undef CPU_SERIAL_IMPLEM //only_for_context cpu_serial
-    #undef CPU_OMP_IMPLEM //only_for_context cpu_openmp
+            part->_num_active_particles = num_active;
+            part->_num_lost_particles = num_lost;
+        }
+
+        #else // not CPU_SERIAL_IMPLEM and not CPU_OMP_IMPLEM
+
+        /*gpufun*/
+        int64_t check_is_active(LocalParticle* part) {
+            return LocalParticle_get_state(part)>0;
+        };
+
+        #endif // CPU_OMP_IMPLEM
+        #endif // CPU_SERIAL_IMPLEM
+
+        #undef CPU_SERIAL_IMPLEM //only_for_context cpu_serial
+        #undef CPU_OMP_IMPLEM //only_for_context cpu_openmp
 
 
-    '''
+        '''
 
         source = '\n\n'.join([src_typedef, src_adders, src_getters,
-                              src_setters, src_scalers, src_exchange,
-                              src_particles_to_local, src_local_to_particles,
-                              src_angles, custom_source])
+                                 src_setters, src_scalers, src_exchange,
+                                 src_particles_to_local, src_local_to_particles,
+                                 src_angles, custom_source])
 
+        source += """
+                    #ifdef XTRACK_GLOBAL_XY_LIMIT
+
+                    /*gpufun*/
+                    void global_aperture_check(LocalParticle* part0) {
+                        //start_per_particle_block (part0->part)
+                            double const x = LocalParticle_get_x(part);
+                            double const y = LocalParticle_get_y(part);
+
+                        int64_t const is_alive = (int64_t)(
+                                  (x >= -XTRACK_GLOBAL_XY_LIMIT) &&
+                                  (x <=  XTRACK_GLOBAL_XY_LIMIT) &&
+                                  (y >= -XTRACK_GLOBAL_XY_LIMIT) &&
+                                  (y <=  XTRACK_GLOBAL_XY_LIMIT) );
+
+                        // I assume that if I am in the function is because
+                            if (!is_alive){
+                               LocalParticle_set_state(part, -1);
+                        }
+                        //end_per_particle_block
+                    }
+
+                    #endif
+
+                    /*gpufun*/
+                    void LocalParticle_add_to_energy(LocalParticle* part, double delta_energy, int pz_only ){
+                        double ptau = LocalParticle_get_ptau(part);
+                        double const p0c = LocalParticle_get_p0c(part);
+
+                        ptau += delta_energy/p0c;
+                        double const old_rpp = LocalParticle_get_rpp(part);
+
+                        LocalParticle_update_ptau(part, ptau);
+
+                        if (!pz_only) {
+                            double const new_rpp = LocalParticle_get_rpp(part);
+                            double const f = old_rpp / new_rpp;
+                            LocalParticle_scale_px(part, f);
+                            LocalParticle_scale_py(part, f);
+                        }
+                    }
+
+
+                    /*gpufun*/
+                    void LocalParticle_update_p0c(LocalParticle* part, double new_p0c_value){
+
+                        double const mass0 = LocalParticle_get_mass0(part);
+                        double const old_p0c = LocalParticle_get_p0c(part);
+                        double const old_delta = LocalParticle_get_delta(part);
+                        double const old_beta0 = LocalParticle_get_beta0(part);
+
+                        double const ppc = old_p0c * old_delta + old_p0c;
+                        double const new_delta = (ppc - new_p0c_value)/new_p0c_value;
+
+                        double const new_energy0 = sqrt(new_p0c_value*new_p0c_value + mass0 * mass0);
+                        double const new_beta0 = new_p0c_value / new_energy0;
+                        double const new_gamma0 = new_energy0 / mass0;
+
+                        LocalParticle_set_p0c(part, new_p0c_value);
+                        LocalParticle_set_gamma0(part, new_gamma0);
+                        LocalParticle_set_beta0(part, new_beta0);
+
+                        LocalParticle_update_delta(part, new_delta);
+
+                        LocalParticle_scale_px(part, old_p0c/new_p0c_value);
+                        LocalParticle_scale_py(part, old_p0c/new_p0c_value);
+
+                        LocalParticle_scale_zeta(part, new_beta0/old_beta0);
+
+                    }
+
+                    /*gpufun*/
+                    void LocalParticle_kill_particle(LocalParticle* part, int64_t kill_state) {
+                        LocalParticle_set_x(part, 1e30);
+                        LocalParticle_set_px(part, 1e30);
+                        LocalParticle_set_y(part, 1e30);
+                        LocalParticle_set_py(part, 1e30);
+                        LocalParticle_set_zeta(part, 1e30);
+                        LocalParticle_update_delta(part, -1);  // zero energy
+                        LocalParticle_set_state(part, kill_state);
+                    }
+                """
         return source
 
     @classmethod
@@ -1730,10 +1813,33 @@ class ParticlesBase(xo.HybridClass):
         self.at_turn = kwargs.get('at_turn', 0)
         self.at_element = kwargs.get('at_element', 0)
         self.weight = kwargs.get('weight', 1)
+        self.x = kwargs.get('x', 0)
+        self.y = kwargs.get('y', 0)
+        self.px = kwargs.get('px', 0)
+        self.py = kwargs.get('py', 0)
         pdg_id = get_pdg_id_from_name(kwargs.get('pdg_id'))
         if not np.isscalar(pdg_id):
             pdg_id = self._context.nparray_to_context_array(pdg_id)
         self.pdg_id = pdg_id
+
+    @classmethod
+    def reference_from_pdg_id(cls, pdg_id, **kwargs):
+        pdg_id = get_pdg_id_from_name(pdg_id)
+        kwargs['pdg_id'] = pdg_id
+        q0 = kwargs.get('q0')
+        mass0 = kwargs.get('mass0')
+        if q0 is None:
+            q, _, _, _ = get_properties_from_pdg_id(pdg_id)
+            kwargs['q0'] = q
+        if mass0 is None:
+            kwargs['mass0'] = get_mass_from_pdg_id(pdg_id)
+
+        particle_ref = cls(**kwargs)
+        if particle_ref._capacity > 1:
+            raise ValueError("The method `reference_from_pdg_id` should have "
+                           + "a `_capacity` of 1. Make sure all properties are "
+                           + "single entries!")
+        return particle_ref
 
     def _allclose(self, a, b, rtol=1e-05, atol=1e-08, mask=None):
         """Substitute for np.allclose that works with all contexts, and
@@ -1972,3 +2078,7 @@ def _mask_to_where(mask, ctx):
     whr = np.where(mask)[0]
     whr = ctx.nparray_to_context_array(whr)
     return whr
+
+
+def reference_from_pdg_id(pdg_id, **kwargs):
+    return Particles.reference_from_pdg_id(pdg_id, **kwargs)
