--- xpart/particles/__init__.py
@@ -1,8 +1,0 @@
-# copyright ############################### #
-# This file is part of the Xpart Package.   #
-# Copyright (c) CERN, 2021.                 #
-# ######################################### #
-
-from .particles import Particles, reference_from_pdg_id
-from .particles_purely_longitudinal import ParticlesPurelyLongitudinal
-from .particles_base import *
